trace record

+++ app/src/main/java/com/boredream/lovebook/utils/TraceUtils.kt
@@ -16,7 +16,7 @@ object TraceUtils {
         return (System.currentTimeMillis() - traceList[0].time) / 1000
     }
 
-    fun calculateDistance(traceList: ArrayList<TraceLocation>): Float {
+    fun calculateDistance(traceList: ArrayList<TraceLocation>): Int {
         var totalDistance = 0f
         for (i in 1 until traceList.size) {
             val distance = AMapUtils.calculateLineDistance(
@@ -25,8 +25,7 @@ object TraceUtils {
             )
             totalDistance += distance
         }
-        // TODO: 应该使用int，但旧数据有float
-        return totalDistance.toInt().toFloat()
+        return totalDistance.toInt()
     }
 
 }
